pyproject.toml
[tool.poetry]
name = "github-pr-fetcher"
version = "1.0.0"
description = "CLI tool for fetching and exporting pull requests from GitHub"
authors = ["Adil Ibraliev <adue@example.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/yourusername/github-pr-fetcher"
repository = "https://github.com/yourusername/github-pr-fetcher"
keywords = ["github", "pull-requests", "cli", "export"]
packages = [{include = "src"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control :: Git",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.8.1"
typer = {extras = ["all"], version = "^0.12.3"}
rich = "^13.7.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
groq = "^0.32.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
gh-pr-fetch = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false